refactor(ecc): move RCX artifact handling into rcx_artifacts module
Per the repository's decomposition rule, path resolution, stale
spef_writer cleanup, and SPEF publication move from the oversized ecc
runner into chipcompiler/tools/ecc/rcx_artifacts.py, with their tests in
test/tools/ecc/test_rcx_artifacts.py; run_rcx keeps orchestrating
extraction.

Publication now also removes the .prev backups after the output set
passes validation, so a successful rerun does not accumulate a second
SPEF set on disk; backup-removal failures are reported without failing
the already-published step. A successful-rerun regression asserts fresh
content and no temp/backup leftovers.

Refs #229
